Add id-pill component
Move the id-fnurgel pill seen in the top right corner of cards to its
own component. Use it in item-local as well.
